requirements.txt
streamlit>=1.28
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
requests>=2.31
google-genai>=1.0
google-play-scraper>=1.2.7
groq>=0.9.0
openai>=1.40.0
